requirements.txt
alpaca-py
pandas
python-dotenv
sqlalchemy
